requirements.txt
pyperclip
requests
PyQt

